# Replace b58-encoded keys with b32 (no padding) + test against multiple ds