--- a/package.json
+++ b/package.json
@@ -82,6 +82,12 @@
       "hash": "QmfVj3x4D6Jkq9SEoi5n2NmoUomLwoeiwnYz2KQa15wRw6",
       "name": "base32",
       "version": "0.0.2"
+    },
+    {
+      "author": "whyrusleeping",
+      "hash": "QmcxDvw8NnJsfdEcfrypwHkLeVxZY2rT8iiWsUuBnw93gb",
+      "name": "go-ds-leveldb",
+      "version": "1.2.0"
     }
   ],
   "gxVersion": "0.4.0",