--- a/pstoreds/keybook.go
+++ b/pstoreds/keybook.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"errors"
 
+	base32 "github.com/whyrusleeping/base32"
+
 	ds "github.com/ipfs/go-datastore"
 	query "github.com/ipfs/go-datastore/query"
 
@@ -13,7 +15,7 @@ import (
 )
 
 // Public and private keys are stored under the following db key pattern:
-// /peers/keys/<b58 of peer id>/{pub, priv}
+// /peers/keys/<b32 peer id no padding>/{pub, priv}
 var (
 	kbBase     = ds.NewKey("/peers/keys")
 	pubSuffix  = ds.NewKey("/pub")
@@ -31,7 +33,7 @@ func NewKeyBook(_ context.Context, store ds.TxnDatastore, _ Options) (pstore.Key
 }
 
 func (kb *dsKeyBook) PubKey(p peer.ID) ic.PubKey {
-	key := kbBase.ChildString(peer.IDB58Encode(p)).Child(pubSuffix)
+	key := kbBase.ChildString(base32.RawStdEncoding.EncodeToString([]byte(p))).Child(pubSuffix)
 
 	var pk ic.PubKey
 	if value, err := kb.ds.Get(key); err == nil {
@@ -68,7 +70,7 @@ func (kb *dsKeyBook) AddPubKey(p peer.ID, pk ic.PubKey) error {
 		return errors.New("peer ID does not match public key")
 	}
 
-	key := kbBase.ChildString(peer.IDB58Encode(p)).Child(pubSuffix)
+	key := kbBase.ChildString(base32.RawStdEncoding.EncodeToString([]byte(p))).Child(pubSuffix)
 	val, err := pk.Bytes()
 	if err != nil {
 		log.Errorf("error while converting pubkey byte string for peer %s: %s\n", p.Pretty(), err)
@@ -82,7 +84,7 @@ func (kb *dsKeyBook) AddPubKey(p peer.ID, pk ic.PubKey) error {
 }
 
 func (kb *dsKeyBook) PrivKey(p peer.ID) ic.PrivKey {
-	key := kbBase.ChildString(peer.IDB58Encode(p)).Child(privSuffix)
+	key := kbBase.ChildString(base32.RawStdEncoding.EncodeToString([]byte(p))).Child(privSuffix)
 	value, err := kb.ds.Get(key)
 	if err != nil {
 		log.Errorf("error while fetching privkey from datastore for peer %s: %s\n", p.Pretty(), err)
@@ -104,7 +106,7 @@ func (kb *dsKeyBook) AddPrivKey(p peer.ID, sk ic.PrivKey) error {
 		return errors.New("peer ID does not match private key")
 	}
 
-	key := kbBase.ChildString(peer.IDB58Encode(p)).Child(privSuffix)
+	key := kbBase.ChildString(base32.RawStdEncoding.EncodeToString([]byte(p))).Child(privSuffix)
 	val, err := sk.Bytes()
 	if err != nil {
 		log.Errorf("error while converting privkey byte string for peer %s: %s\n", p.Pretty(), err)