Resolves #38.

Also adds support for running tests against multiple ds implementations, albeit disabled as go-ds-leveldb doesn't support TTL (for now).

Very interesting benchmark improvements, especially on get (look at the cacheful get):

```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkDsPeerstore/Badger/AddGetAndClearAddrs-1Addrs-Caching-8           1007578       1299832       +29.01%
BenchmarkDsPeerstore/Badger/AddGetAndClearAddrs-10Addrs-Caching-8          4184801       3579472       -14.46%
BenchmarkDsPeerstore/Badger/AddGetAndClearAddrs-100Addrs-Caching-8         12334153      10150631      -17.70%
BenchmarkDsPeerstore/Badger/Get1000PeersWithAddrs-1Addrs-Caching-8         5566697       4085272       -26.61%
BenchmarkDsPeerstore/Badger/Get1000PeersWithAddrs-10Addrs-Caching-8        18989271      17957917      -5.43%
BenchmarkDsPeerstore/Badger/Get1000PeersWithAddrs-100Addrs-Caching-8       174546766     166842294     -4.41%
BenchmarkDsPeerstore/Badger/AddAddrs-1Addrs-Caching-8                      123719        130628        +5.58%
BenchmarkDsPeerstore/Badger/AddAddrs-10Addrs-Caching-8                     252290        254833        +1.01%
BenchmarkDsPeerstore/Badger/AddAddrs-100Addrs-Caching-8                    763895        773458        +1.25%
BenchmarkDsPeerstore/Badger/SetAddrs-1Addrs-Caching-8                      125053        122690        -1.89%
BenchmarkDsPeerstore/Badger/SetAddrs-10Addrs-Caching-8                     251483        250387        -0.44%
BenchmarkDsPeerstore/Badger/SetAddrs-100Addrs-Caching-8                    772552        779483        +0.90%
BenchmarkDsPeerstore/Badger/GetAddrs-1Addrs-Caching-8                      1694          654           -61.39%
BenchmarkDsPeerstore/Badger/GetAddrs-10Addrs-Caching-8                     1681          739           -56.04%
BenchmarkDsPeerstore/Badger/GetAddrs-100Addrs-Caching-8                    2107          1112          -47.22%
BenchmarkDsPeerstore/Badger/AddAddrs-1Addrs-Cacheless-8                    108746        110216        +1.35%
BenchmarkDsPeerstore/Badger/AddAddrs-10Addrs-Cacheless-8                   251620        251198        -0.17%
BenchmarkDsPeerstore/Badger/AddAddrs-100Addrs-Cacheless-8                  768939        770448        +0.20%
BenchmarkDsPeerstore/Badger/SetAddrs-1Addrs-Cacheless-8                    135093        115331        -14.63%
BenchmarkDsPeerstore/Badger/SetAddrs-10Addrs-Cacheless-8                   249202        252054        +1.14%
BenchmarkDsPeerstore/Badger/SetAddrs-100Addrs-Cacheless-8                  787850        762799        -3.18%
BenchmarkDsPeerstore/Badger/GetAddrs-1Addrs-Cacheless-8                    21903         23058         +5.27%
BenchmarkDsPeerstore/Badger/GetAddrs-10Addrs-Cacheless-8                   86672         80272         -7.38%
BenchmarkDsPeerstore/Badger/GetAddrs-100Addrs-Cacheless-8                  510136        492857        -3.39%
BenchmarkDsPeerstore/Badger/AddGetAndClearAddrs-1Addrs-Cacheless-8         1617704       1441082       -10.92%
BenchmarkDsPeerstore/Badger/AddGetAndClearAddrs-10Addrs-Cacheless-8        7408843       6074805       -18.01%
BenchmarkDsPeerstore/Badger/AddGetAndClearAddrs-100Addrs-Cacheless-8       17678520      13696671      -22.52%
BenchmarkDsPeerstore/Badger/Get1000PeersWithAddrs-1Addrs-Cacheless-8       5540977       4033391       -27.21%
BenchmarkDsPeerstore/Badger/Get1000PeersWithAddrs-10Addrs-Cacheless-8      18949443      17345810      -8.46%
BenchmarkDsPeerstore/Badger/Get1000PeersWithAddrs-100Addrs-Cacheless-8     176095545     167676240     -4.78%
```