@raulk 2018-10-04
@bigs I'd really like to keep the structural changes made in tests to test against multiple datastores.

Even if we can't enable the leveldb ones due to the TTL thing (might just implement it tomorrow), they were very useful to get the metrics here against a previous commit: https://github.com/libp2p/go-libp2p-peerstore/issues/31#issuecomment-426735525

I don't think the TODO hurts us.