@bigs 2018-10-04
sounds good!

On Thu, Oct 4, 2018 at 17:40 Raúl Kripalani <notifications@github.com>
wrote:

> @bigs <https://github.com/bigs> I'd really like to keep the structural
> changes made in tests to test against multiple datastores.
>
> Even if we can't enable the leveldb ones due to the TTL thing (might just
> implement it tomorrow), they were very useful to get the metrics here
> against a previous commit: #31 (comment)
> <https://github.com/libp2p/go-libp2p-peerstore/issues/31#issuecomment-426735525>
>
> I don't think the TODO hurts us.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/libp2p/go-libp2p-peerstore/pull/40#issuecomment-427178898>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AANBWgoB15qfEB_MD4riHggPM0xSosBDks5uhoA4gaJpZM4XHx2z>
> .
>